backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
bcrypt==4.1.2
google-generativeai==0.8.3
Werkzeug==3.0.1 